// src/app/model/ordencompra.ts
export class OrdenCompraBuscar {
    nroordencompra: number;
    estado: string;
    tipoorden: string;
    comprador: string;
    proveedor: string;
    empresacompradora?: string;
    atenciona: string;
    version: string;
    total: string;
    fecharegistro: string;

}
export class OrdenCompraFiltros {
    nroordencompra?: string;
    estado?: string;
    material?: boolean;
    servicio?: boolean;
    fechacreacioninicio?: Date;
    fechacreacionfin?: Date;
    constructor() {
        this.fechacreacioninicio = new Date();
        this.fechacreacionfin = new Date();
    }
}